# server/http/handlers/openapi.yaml
openapi: 3.0.3
info:
  title: Keyorix API
  version: 1.0.0
  description: |
    Keyorix is a self-hostable secrets manager with scoped RBAC, audit logging,
    rotation, sharing, and machine-identity authentication.

    ## Authentication
    Most endpoints require a Bearer token: `Authorization: Bearer <token>`.
    Three credential types are accepted, all via the same header:
      - **Session tokens** — issued by `POST /auth/login` (and `/auth/refresh`).
      - **Personal access tokens** (`kx_pat_…`) — created at `POST /api/v1/auth/tokens`.
      - **Machine tokens** (`kx_machine_…`, ADR-030) — issued per machine identity.
    Public endpoints (login, refresh, logout, password-reset, system init, the
    setup-token flow, health, and Prometheus metrics) need no token.

    ## Response envelope
    Unless noted, success responses wrap the payload as `{ "data": <…>, "message": "<…>" }`.
    Errors use `{ "error": { "code": <int>, "message": <string>, "type": <string> } }`.

    ## Authorization
    Operations are gated by scoped RBAC permissions (e.g. `secrets.read`,
    `roles.assign`) at a project/environment scope. A 403 indicates the caller's
    roles do not grant the permission at the target scope.
servers:
  - url: /
    description: Same-origin (paths are absolute from the host root)
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/risk-exceptions:
    get:
      tags: [Dashboard]
      summary: Risk register — governed, time-bound exceptions (ISO 27001 A.5.8)
      description: >-
        The risk register: governed acceptances of a known control gap, each with an
        owner, justification, and sunset date. Active only by default; ?all=true
        includes expired exceptions (history). Each carries a computed status
        (active | expired | revoked). Requires system.read.
      operationId: listRiskExceptions
      parameters:
        - {name: all, in: query, schema: {type: boolean}, description: "Include expired exceptions."}
      security: [{bearerAuth: []}]
      responses:
        '200': {description: "Envelope {data}. data: {exceptions[] {id, title, category, reference, justification, status, expires_at, created_by, created_at}}."}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
    post:
      tags: [Dashboard]
      summary: Record a risk exception (accept a control gap for a bounded time)
      description: "Accept a known gap with a justification and a required future expiry. Requires system.write."
      operationId: createRiskException
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, justification, expires_at]
              properties:
                title: {type: string}
                category: {type: string, enum: [sod, mfa, rotation, dormant_access, classification, other]}
                reference: {type: string, description: "What it applies to (a user, an SoD pair, a secret)."}
                justification: {type: string}
                expires_at: {type: string, format: date-time, description: "RFC3339; must be in the future."}
      responses:
        '201': {description: "Envelope {data, message}. data: {exception}."}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
  /api/v1/risk-exceptions/{id}:
    delete:
      tags: [Dashboard]
      summary: Revoke a risk exception before its expiry (ISO 27001 A.5.8)
      operationId: revokeRiskException
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer}}
      security: [{bearerAuth: []}]
      responses:
        '200': {description: "Envelope {data, message}. Risk exception revoked."}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Session token, personal access token (kx_pat_…), or machine token (kx_machine_…).
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code: {type: integer}
                  message: {type: string}
                  type: {type: string}
...